stark-interface-main/J.A.R.V.I.S.-main/J.A.R.V.I.S.-main/requirements.txt
customtkinter==5.2.2
dateparser==1.2.1
gTTS==2.5.4
langdetect==1.0.9
numpy==1.26.4
opencv_contrib_python==4.11.0.86
opencv_python==4.9.0.80
playsound==1.2.2
pygame==2.6.1
protobuf==5.29.4
PySide6==6.9.1
PySide6==6.10.0
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dotenv==1.2.1
pyttsx3==2.98
Requests==2.32.5
speechrecognition==3.14.2
